break out ApplicationThresholdSerializer
