+++ api/serializers.py
@@ -30,15 +30,6 @@ from threepio import logger
 
 
 # Serializers
-class ApplicationThresholdSerializer(serializers.ModelSerializer):
-    """
-    """
-    min_ram= serializers.IntegerField(source="memory_min")
-    min_disk = serializers.IntegerField(source="storage_min")
-    class Meta:
-        model = ApplicationThreshold
-        exclude = ('id', 'application', 'memory_min', 'storage_min')
-
 class ApplicationSerializer(serializers.ModelSerializer):
     """
     """
