+++ api/serializers/__init__.py
@@ -10,3 +10,4 @@ from .account_serializer import AccountSerializer
 from .provider_serializer import ProviderSerializer
 from .cleaned_identity_serializer import CleanedIdentitySerializer
 from .boot_script_serializer import BootScriptSerializer
+from .application_threshold_serializer import ApplicationThresholdSerializer
